refactor: separate Base and EVM providers logic
