+++ .prettierrc.js
@@ -4,4 +4,5 @@ module.exports = {
   bracketSpacing: false,
   singleQuote: true,
   trailingComma: 'all',
+  plugins: ["prettier-plugin-sort-json"]
 };
