+++ package.json
@@ -53,7 +53,6 @@
     "@types/node": "^20.1.2",
     "@types/react-native": "^0.71.12",
     "jest": "^29.5.0",
-    "prettier": "^2.8.8",
     "react": "18.2.0",
     "react-native": "0.71.12",
     "react-native-ble-plx": "^2.0.3",
@@ -63,7 +62,16 @@
     "tslint-config-prettier": "^1.18.0",
     "typedoc": "^0.24.7",
     "typedoc-plugin-markdown": "^3.15.3",
-    "typescript": "^5.0.4"
+    "typescript": "^5.0.4",
+    "@typescript-eslint/eslint-plugin": "6.0.0",
+    "@typescript-eslint/parser": "6.0.0",
+    "eslint": "8.45.0",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-jest": "27.4.2",
+    "eslint-plugin-prettier": "5.0.0",
+    "eslint-plugin-react-native": "4.0.0",
+    "prettier": "3.0.0",
+    "prettier-plugin-sort-json": "4.0.0"
   },
   "react-native": {
     "crypto": "react-native-crypto",
