+++ src/index.ts
@@ -1,5 +1,5 @@
+import * as constants from './constants';
 import * as providers from './providers';
 import * as utils from './utils';
-import * as constants from './constants';
 
 export {providers, utils, constants};
