+++ src/providers/hot/index.ts
@@ -1,2 +1,2 @@
-export * from './provider';
+export * from './evm-provider';
 export * from './types';
