+++ src/providers/hot/types.ts
@@ -1,4 +1,4 @@
-export type ProviderHotOptions = {
+export type ProviderHotBaseOptions = {
   account: string;
   getPassword: () => Promise<string>;
 };
